requirements.txt
Flask
Flask-SocketIO
pycryptodome
eventlet
